move region2D, fix names
move region2D to Game
