+++ Game/Makefile
@@ -2,7 +2,8 @@ all: ../$(OBJ_DIR)/Game/enemy.o \
 ../$(OBJ_DIR)/Game/player.o \
 ../$(OBJ_DIR)/Game/runner.o \
 ../$(OBJ_DIR)/Game/scene.o \
-../$(OBJ_DIR)/Game/weapon.o
+../$(OBJ_DIR)/Game/weapon.o \
+../$(OBJ_DIR)/Game/region2D.o \
 
 ../$(OBJ_DIR)/Game/enemy.o : enemy.c
 	@$(CC) $(CFLAGS) -c $^ -o $@
@@ -17,4 +18,7 @@ all: ../$(OBJ_DIR)/Game/enemy.o \
 	@$(CC) $(CFLAGS) -c $^ -o $@
 
 ../$(OBJ_DIR)/Game/weapon.o : weapon.c
+	@$(CC) $(CFLAGS) -c $^ -o $@
+
+../$(OBJ_DIR)/Game/region2D.o : region2D.c
 	@$(CC) $(CFLAGS) -c $^ -o $@
