+++ Game/game.h
@@ -3,3 +3,4 @@
 #include "enemy.h"
 #include "scene.h"
 #include "weapon.h"
+#include "region2D.h"
