+++ op_engine/op_engine.h
@@ -8,4 +8,3 @@
 #include "object.h"
 #include "mesh.h"
 #include "input.h"
-#include "region2D.h"
